+++ actix-router/src/resource_path.rs
@@ -27,7 +27,7 @@ impl<'a> ResourcePath for &'a str {
 
 impl ResourcePath for bytestring::ByteString {
     fn path(&self) -> &str {
-        &*self
+        self
     }
 }